move MemoryData to memory.rs; remove all the Hashing stuff
